Dirs restructured so src is beside Code::Blocks project.
