+++ .gitignore
@@ -1,0 +1,22 @@
+# unit test runners e.g run-test-ascii_num.c.
+run-*.c
+*_Runner.c
+# unit test coverage files.
+*.gcda
+*.gcno
+# gcovr results (for now)
+gcovr*.html
+# in projects
+*.o
+*.exe
+# backups from e.g Winmerge
+*.bak
+# libs
+*.a
+#copies
+* - Copy.*
+# Code::Blocks
+codeblocks_gcc/*.depend
+codeblocks_gcc/*.layout
+# codeblocks_gcc/main.c
+
