--- libs_support.h
@@ -1,11 +1,0 @@
-#ifndef LIBS_SUPPORT_H_INCLUDED
-#define LIBS_SUPPORT_H_INCLUDED
-
-#include "GenericTypeDefs.h"
-#define RAM_IS
-#define BOOLEAN BOOL
-#define bool BOOL
-
-#define REGEXLT_PRINT_STDIO
-
-#endif // LIBS_SUPPORT_H_INCLUDED
