+++ src/regexlt_right_op.c
